requirements.txt
hdx-python-api
fiona
requests
